Move kanban_board CRUD operations to a separate file
